alumni-tracer-system/backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
getenv==0.2.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
zope.interface==8.1.1
